Update README.md
Since version 1.2, airkick no longer consists of just an executable but also a textfile
containing vendor OUIs. This textfile is installed at a location depending on the autotools
installation prefix, a variable meant to be customizable. Thus it wouldn't make much
sense to force a particular directory layout on the user with a prebuilt binary.